[errorTrace] - renamed module to match camelcase spec (made docker UI crash)
