+++ server/src/processing/HeadingDetectionDtModule/HeadingDetectionDtModule.ts
