+++ server/src/processing/HeadingDetectionDtModule/defaultConfig.json
