+++ server/src/processing/HeadingDetectionDtModule/model.js
